requirements.txt
requests
pbkdf2
pycryptodomex
python-dotenv